feat: allow overriding actor app name (#69)
This PR adds the feature to override the default app actor name with an environment variable APP_ACTOR_NAME_OVERRIDE which used to be hardcoded to "approveman[bot]".

This is useful for GitHub Enterprise users whose app content name is different from the GitHub marketplace. Before the change, GitHub Enterprise users will have trouble dismissing outdated approvals since the name it searches for is "approveman[bot]" (for details on this problem, see #64 ).

This PR also adds a check so that if a GitHub Enterprise user (inferred from GHE_HOST) deploys the app without setting APP_ACTOR_NAME_OVERRIDE will post a failing status upon any event processed.

This is not ideal because we want to fail the moment the app starts running instead of when an event is received. However, Probot currently doesn't support this kind of health check as far as I know. I have opened an issue for Probot to add this capability into their infra (for details and progress, see probot/probot#1340 ).