feat: add global blacklisting capability (#156)
This PR adds the capability for repository owners to set a list of globally blacklisted usernames. This can be useful when certain users are spotted to abuse the automatic approval process and checks in unwanted content that violates policies set by the owner. In the future, a whitelisting capability is planned for people who want to keep the authorized user within the team.

* feat: add logic to globally block users
* docs: update config docs
* fix: optimize imports