feat: Support user whitelist (opposite of current blacklist) (#494)
Prior to this change, users could be explicitly disallowed, providing
for the case where most users with access to the repository could
benefit from auto-approval, but not all.  However, there are at least a
couple of use-cases where many people have access to a repo, with the
specific set of people changing over time, but only a limited and
consistent set of users should have their PRs auto approved. Such cases
include:
* Sub-admin super-users who are permitted to make changes to specific
  files without oversight, but with repository checks still applied.
* Machine Users that commit back to the repository as part of a CI/CD
  pipeline without human oversight in the general case; repository
  checks and directory matching rules are used to detect bad actors and
  leave PRs unmerged for human oversight.

This change adds a new optional configuration field to support the
specification of explicitly allowed users.  When not provided or left
empty, the existing behavior prevails: All users are allowed, unless
blacklisted.  When provided, on the other hand, only the given users are
allowed (effectively blacklisting everyone _else_).  Finally, were the
same user is specified in both the new configuration and a blacklist,
the blacklist takes precedence, and the user is blocked.

Co-authored-by: Tianhao Zhou <tianhaoz@umich.edu>